experimental-plugins/agentic-workflows/pyproject.toml
[project]
name = "agentic-workflows"
version = "0.1.0"
description = "Agentic workflow orchestration for Claude Code"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "filelock>=3.12",
]

[project.scripts]
agentic-workflow = "agentic_workflows.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_workflows"]
